refactor: skip step 3 if lua template
